refactor(scripts): FILENAME_TO_SCHEMA を schema_mapping.py に集約（#172）
retrospective 2026-05-01 Try T-6 への対応。validate-schemas.py / eval-runner.py
で重複していた 16 entry mapping を 1 箇所に集約、新 schema 追加時の同期漏れ
リスクを構造的に解消する。

- scripts/schema_mapping.py: 新規（FILENAME_TO_SCHEMA + SCHEMAS_DIR + lookup_schema）
- scripts/validate-schemas.py: 重複削除 + sys.path.insert + import
- scripts/eval-runner.py: 重複削除 + import + EVAL_RUNNER_VERSION 1.0.0 → 1.1.0
- tests/run-tests.sh: 21 件 PASS 維持（後方互換）

closes #172
