+++ docs/working/TASK-0051/todo.md
@@ -1,0 +1,13 @@
+# EXECUTION TODO: TASK-0051 / Issue #172
+
+> Mode: **light**
+
+- [x] PBI INPUT
+- [x] plan + test-cases
+- [x] scripts/schema_mapping.py 新設
+- [x] scripts/validate-schemas.py を import に置換
+- [x] scripts/eval-runner.py を import に置換 + version bump
+- [x] sh tests/run-tests.sh で 21 PASS
+- [x] handoff.md
+- [x] commit + push + PR
+- [ ] CI 緑確認 → C-4
